Remove inaccurate/redundant info, minor changes
Not sure what "All major universities in Singapore" refers to here. We're open to all students. Deleted it.
The "also welcome" section lists things that are already part of the sponsorship tiers. Listing it separately looks a bit confusing. Deleted it.
Changed the wording of a few sentences a little bit.